Partly fix art with sabermod 4.8
First error fixed - now there's another error:
F/art     (  604): art/runtime/well_known_classes.cc:118] Couldn't find method "nativeLoad" with signature "(Ljava/lang/String;Ljava/lang/ClassLoader;Ljava/lang/String;)Ljava/lang/String;"
But I have no idea how to fix that. I never used JNI and therefore don't know how it works.
The signature seems to be right as can be seen in runtime/native/java_lang_Runtime.cc => search for "NATIVE_METHOD".

+++ runtime/mirror/art_method.cc
@@ -84,7 +84,7 @@ void ArtMethod::SetDexCacheInitializedStaticStorage(ObjectArray<StaticStorageBas
       new_value, false);
 }
 
-size_t ArtMethod::NumArgRegisters(const StringPiece& shorty) {
+uint32_t ArtMethod::NumArgRegisters(const StringPiece& shorty) {
   CHECK_LE(1, shorty.length());
   uint32_t num_registers = 0;
   for (int i = 1; i < shorty.length(); ++i) {
